+++ library/src/main/java/com/matthewtamlin/java_compiler_utilities/collectors/ElementCollector.java
@@ -1,4 +1,4 @@
-package com.matthewtamlin.java_compiler_utilities.element_util.collectors;
+package com.matthewtamlin.java_compiler_utilities.collectors;
 
 import javax.annotation.processing.AbstractProcessor;
 import javax.lang.model.element.Element;
