+++ library/src/main/java/com/matthewtamlin/java_compiler_utilities/element_supplier/ElementId.java
@@ -1,4 +1,4 @@
-package com.matthewtamlin.java_compiler_utilities.element_util.element_supplier;
+package com.matthewtamlin.java_compiler_utilities.element_supplier;
 
 /**
  * Defines an ID for an element in a source file.
