+++ library/src/main/java/com/matthewtamlin/java_compiler_utilities/in_memory_file_utils/InMemoryJavaFileManager.java
@@ -1,4 +1,4 @@
-package com.matthewtamlin.java_compiler_utilities.element_util.in_memory_file_utils;
+package com.matthewtamlin.java_compiler_utilities.in_memory_file_utils;
 
 import javax.tools.FileObject;
 import javax.tools.ForwardingJavaFileManager;
