+++ library/src/main/java/com/matthewtamlin/java_compiler_utilities/in_memory_file_utils/InMemoryJavaFileObject.java
@@ -1,4 +1,4 @@
-package com.matthewtamlin.java_compiler_utilities.element_util.in_memory_file_utils;
+package com.matthewtamlin.java_compiler_utilities.in_memory_file_utils;
 
 import com.google.common.io.ByteSource;
