+++ library/src/test/java/com/matthewtamlin/java_compiler_utilities/Tag1.java
@@ -1,7 +1,7 @@
-package com.matthewtamlin.java_compiler_utilities.element_util;
+package com.matthewtamlin.java_compiler_utilities;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.SOURCE)
-public @interface Tag2 {}
+public @interface Tag1 {}
