+++ library/src/test/java/com/matthewtamlin/java_compiler_utilities/TestElementSupplier.java
@@ -1,10 +1,10 @@
-package com.matthewtamlin.java_compiler_utilities.element_util;
+package com.matthewtamlin.java_compiler_utilities;
 
 import com.google.common.collect.ImmutableSet;
 import com.google.testing.compile.JavaFileObjects;
-import com.matthewtamlin.java_compiler_utilities.element_util.element_supplier.CompilerMissingException;
+import com.matthewtamlin.java_compiler_utilities.element_supplier.CompilerMissingException;
 import com.matthewtamlin.java_compiler_utilities.element_util.element_supplier.ElementSupplier;
-import com.matthewtamlin.java_compiler_utilities.element_util.element_supplier.UniqueElementNotFoundException;
+import com.matthewtamlin.java_compiler_utilities.element_supplier.UniqueElementNotFoundException;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
